+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 - Use Lua table for Metadata field when updating a user.
 - Updated configuration variable names. Most importantly `DB` is now changed to `database.address`.
 - Moved all `nakamax` functions into `nakama`. 
+- Invalid config file, or invalid command line config option prevents server from starting.
 
 ## [1.0.0-rc.1] - 2017-07-18
 ### Added